# sql: don't re-decode index keys when unnecessary