Previously, in multi column family tables, the row fetcher would
unconditionally decode the prefix of every key, then check to see if
that prefix was identical to the last prefix to determine whether or not
a new row was found.

This commit reverses the order - first we check the prefix, so that if
it's identical to the last prefix, we can completely skip decoding the
key.

Fixes #21565.

```
name                     old time/op    new time/op    delta
CountTwoCF/Cockroach-24    51.3ms ± 1%    50.4ms ± 1%  -1.70%  (p=0.000 n=10+9)

name                     old alloc/op   new alloc/op   delta
CountTwoCF/Cockroach-24    6.97MB ± 0%    6.98MB ± 0%    ~     (p=0.515 n=10+8)

name                     old allocs/op  new allocs/op  delta
CountTwoCF/Cockroach-24     1.99k ± 2%     1.99k ± 1%    ~     (p=0.828 n=10+8)
```

I'm not sure why the benchmark results don't look better. In local
testing, I see more like a 10% improvement. Either way, this is an
unambiguous win.

Release note: None