@jordanlewis 2018-10-22
Reason I found this, by the way, was trying to fix up the column rowfetcher from #31354 - this isn't completely out of the blue :)