@jordanlewis 2018-10-22
Ooh, looks like this fixes an ancient issue I filed: #21565.